Remove secure erase feature
This method of secure erase is only useful for conventional HDDs, which
are not commonly used as OS disks these days. SSDs are also notoriously
nuanced when it comes to erasing, and to do this programatically (and
reliably) would be prohibitively difficult.

To prevent confusion, avoid giving users a false sense of security, and
reduce maintenance burden, I've decided to remove the feature. If a user
would like to erase their disk, they should do so manually.